+++ database/migrations/2020_12_02_112350_add_references_to_article_translations.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddReferencesToArticleTranslations extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('article_translations', function (Blueprint $table) {
+            $table->string('references')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('article_translations', function (Blueprint $table) {
+            $table->dropColumn('references');
+        });
+    }
+}