+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file[^1].
 - Use laravel-mix for (some) JavaScript files
 - Add item (artwork) data to GTM dataLayer
 - Use atymic/deployer-php-action@2.0 for auto-deploys
+- Add block to article / collection to list references 
 
 ## [2.11.1] - 2020-11-11
 ### Changed